Documents/mhtcet/backend/requirements.txt
fastapi
uvicorn
pandas
tabula-py
google-generativeai
python-multipart
pdfplumber 